feat: extracts fetDataSets, converts to query, also refactors inline styling
