+++ src/components/syncs/SyncDataSet.js
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from "react";
+import { useQuery } from "react-query";
 import MUIDataTable from "mui-datatables";
 import PeriodPicker from "../shared/PeriodPicker";
 import PluginRegistry from "../core/PluginRegistry";
@@ -8,104 +9,43 @@ import { constructDataSyncTableColumns } from "./tables";
 import _ from "lodash";
 import { Button, Typography, makeStyles, Paper, CircularProgress } from "@material-ui/core";
 import { useTranslation } from "react-i18next";
+import { fetchDataSets } from "./fetchDataSets";
 
 const useStyles = makeStyles({
   aligned: {
     textAlign: "center",
     verticalAlign: "middle",
   },
+  root: { minHeight: "85vh", paddingTop: "14px", paddingLeft: "14px" },
+  header: {
+    display: "flex",
+    flexDirection: "row",
+    alignContent: "center",
+    justifyContent: "space-between",
+    paddingTop: "5px",
+    paddingLeft: "5px",
+    marginBottom: "20px",
+  },
+  headerTitleHolder: { display: "inline-flex" },
+  headerTitle: { marginRight: "20px" },
+  picker: { padding: "1px" },
+  syncButton: { float: "right" },
 });
 
 const SyncDataSet = (props) => {
+  const classes = useStyles();
   const { t } = useTranslation();
   const dhis2 = PluginRegistry.extension("core.dhis2");
   const dhis2RootUrl = dhis2.baseUrl;
-
-  const project = PluginRegistry.extension("hesabu.project");
   const DataEntries = PluginRegistry.extension("dataentry.dataEntries");
-  const classes = useStyles(props);
   const allDataEntries = DataEntries.getAllDataEntries();
-  const [loading, setLoading] = useState(undefined);
   const [loadingStatus, setLoadingStatus] = useState(undefined);
-  const [dataElementsById, setDataElementsById] = useState(undefined);
-  const [dataSetsById, setDataSetsById] = useState(undefined);
-  const [contractsByDataEntryCode, setContractsByDataEntryCode] = useState(undefined);
 
   const period = props.match.params.period;
+  const fetchDataSetsQuery = useQuery(["dataSets", period], () => fetchDataSets(allDataEntries, period));
 
-  const fetchDataSets = async () => {
-    setLoading(true);
-    const api = await dhis2.api();
-    const dataElements = _.keyBy(
-      (await api.get("dataElements", { paging: false, fields: "id,name" })).dataElements,
-      (de) => de.id,
-    );
-    setDataElementsById(dataElements);
-    const dataSetIds = allDataEntries.flatMap((d) => (d.dataSetId ? [d.dataSetId] : d.dataSetIds));
-    const ds = await api.get("dataSets", {
-      filter: ["id:in:[" + dataSetIds.join(",") + "]"],
-      paging: false,
-      fields: ":all,dataSetElements[dataElement[id,name]]organisationUnits[id,name],workflow[:all]",
-    });
-    const dataSetsById = _.keyBy(ds.dataSets, (d) => d.id);
-    setDataSetsById(dataSetsById);
-
-    const contractService = PluginRegistry.extension("contracts.service");
-    const contracts = await contractService.findAll();
-
-    const activeCotnracts = contracts.filter((contract) => contract.matchPeriod(period));
-    const contractsByDataEntryCode = {};
-    for (let dataEntry of allDataEntries) {
-      const dataEntryContracts = activeCotnracts.filter((activeContract) =>
-        dataEntry.contracts.some((contractFilter) =>
-          contractFilter.every((code) => activeContract.codes.includes(code)),
-        ),
-      );
-      const dataSets = dataEntry.dataSetId
-        ? [dataSetsById[dataEntry.dataSetId]]
-        : dataEntry.dataSetIds
-        ? dataEntry.dataSetIds.map((id) => dataSetsById[id])
-        : [];
-      const results = [];
-      for (let dataSet of dataSets) {
-        contractsByDataEntryCode[dataEntry.code];
-        const dataSetOrgunits = new Set(dataSet.organisationUnits.map((ou) => ou.id));
-        const missingOrgunits = dataEntryContracts.map((c) => c.orgUnit).filter((ou) => !dataSetOrgunits.has(ou.id));
-
-        let expectedDataElements = [];
-        const missingDataElements = [];
-        if (dataEntry.hesabuInputs) {
-          const project_descriptor = project(period);
-          const payment = project_descriptor.payment_rules[dataEntry.hesabuPayment];
-          const hesabuPackage = payment.packages[dataEntry.hesabuPackage];
-          expectedDataElements = hesabuPackage.activities
-            .flatMap((activity) => dataEntry.hesabuInputs.map((state) => activity[state]))
-            .filter((de) => de);
-          const dataSetElements = new Set(dataSet.dataSetElements.map((dse) => dse.dataElement.id));
-          for (let expectedDE of expectedDataElements) {
-            if (expectedDE && !dataSetElements.has(expectedDE)) {
-              missingDataElements.push(expectedDE);
-            }
-          }
-        }
-
-        results.push({
-          dataSet: dataSet,
-          activeContracts: dataEntryContracts,
-          missingOrgunits: missingOrgunits,
-          expectedDataElements: expectedDataElements,
-          missingDataElements: missingDataElements,
-        });
-      }
-      contractsByDataEntryCode[dataEntry.code] = results;
-    }
-    setContractsByDataEntryCode(contractsByDataEntryCode);
-    setLoading(false);
-  };
-
-  useEffect(() => {
-    fetchDataSets();
-  }, []);
+  const dataElementsById = fetchDataSetsQuery?.data?.dataElementsById;
+  const contractsByDataEntryCode = fetchDataSetsQuery?.data?.contractsByDataEntryCode;
 
   const updateOu = async (myDataSet, missingOrgunits) => {
     setLoadingStatus(`Updating ${myDataSet.name}`);
@@ -123,29 +63,32 @@ const SyncDataSet = (props) => {
     await api.update("dataSets/" + dataSet.id, dataSet);
   };
 
-  const addAllMissingOus = async () => {
-    setLoading(true);
-    const contracts = Object.values(contractsByDataEntryCode);
-    for (let contractGroups of contracts) {
-      for (let contract of contractGroups) {
-        if (contract.missingOrgunits.length > 0) {
-          await updateOu(contract.dataSet, contract.missingOrgunits);
+  const addAllMissingOusQuery = useQuery(
+    ["addAllMissingOrgunits"],
+    async () => {
+      const contracts = Object.values(contractsByDataEntryCode);
+      for (let contractGroups of contracts) {
+        for (let contract of contractGroups) {
+          if (contract.missingOrgunits.length > 0) {
+            await updateOu(contract.dataSet, contract.missingOrgunits);
+          }
         }
       }
-    }
-    setLoading(false);
-    fetchDataSets();
-  };
+      fetchDataSetsQuery.refetch();
+    },
+    {
+      enabled: false,
+    },
+  );
+
+  const loading = fetchDataSetsQuery.isLoading || addAllMissingOusQuery.isLoading;
 
   const addSingleMissingOu = async (myDataSet, missingOrgunits) => {
-    setLoading(true);
     await updateOu(myDataSet, missingOrgunits);
-    setLoading(false);
-    fetchDataSets();
+    fetchDataSetsQuery.refetch();
   };
 
   const addMissingDe = async (dataEntry) => {
-    setLoading(true);
     const missing = contractsByDataEntryCode[dataEntry.code][0].missingDataElements;
     const api = await dhis2.api();
     const dataSet = await api.get("dataSets/" + dataEntry.dataSetId, {
@@ -166,8 +109,7 @@ const SyncDataSet = (props) => {
     } catch (error) {
       alert("Something went wrong was not able to update the dataset " + JSON.stringify(error));
     }
-    setLoading(false);
-    fetchDataSets();
+    fetchDataSetsQuery.refetch();
   };
 
   const data = allDataEntries.map((dataEntry) => {
@@ -198,43 +140,35 @@ const SyncDataSet = (props) => {
   });
 
   return (
-    <Paper style={{ minHeight: "85vh", paddingTop: "14px", paddingLeft: "14px" }}>
+    <Paper className={classes.root}>
       <div>
-        <div
-          style={{
-            display: "flex",
-            flexDirection: "row",
-            alignContent: "center",
-            justifyContent: "flex-start",
-            paddingTop: "5px",
-            paddingLeft: "5px",
-            marginBottom: "20px",
-          }}
-        >
-          <Typography variant="h6" style={{ marginRight: "20px" }}>
-            Dataset Synchronisation for
-          </Typography>
-          <div style={{ padding: "1px" }}>
-            <PeriodPicker
-              disableInputLabel={true}
-              period={period}
-              periodDelta={{
-                before: 5,
-                after: 5,
-              }}
-              onPeriodChange={(newPeriod) => {
-                props.history.push("/sync/datasets/" + newPeriod);
-              }}
-            />
+        <div className={classes.header}>
+          <div className={classes.headerTitleHolder}>
+            <Typography variant="h6" className={classes.headerTitle}>
+              Dataset Synchronisation for
+            </Typography>
+            <div classes={classes.picker}>
+              <PeriodPicker
+                disableInputLabel={true}
+                period={period}
+                periodDelta={{
+                  before: 5,
+                  after: 5,
+                }}
+                onPeriodChange={(newPeriod) => {
+                  props.history.push("/sync/datasets/" + newPeriod);
+                }}
+              />
+            </div>
+          </div>
+          <div className={classes.syncButton}>
+            <Button onClick={addAllMissingOusQuery.refetch} color="primary">
+              Synchronize all {loading && loadingStatus ? <CircularProgress size={15} /> : ""}
+            </Button>
           </div>
         </div>
       </div>
-      <div style={{ float: "right" }}>
-        <Button onClick={addAllMissingOus} color="primary">
-          Synchronize all {loading && loadingStatus ? <CircularProgress size={15} /> : ""}
-        </Button>
-      </div>
-      <div style={{ paddingTop: "50px" }}>
+      <div>
         <MUIDataTable data={data} columns={columns} options={options} />
       </div>
     </Paper>
